requirements.txt
asgiref==3.7.2
Django==4.2.3
django-extensions==3.2.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
DateTime==4.7
docx2pdf==0.1.8
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
pdf2image==1.16.3
Pillow==10.0.0
pygsheets==2.0.6
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyMuPDF==1.22.5
PyPDF2==3.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.28.1